02-functions/code/upload_url/requirements.txt
google-cloud-storage>=2.10.0
google-auth>=2.22.0
